requirements.txt
# Core
anthropic>=0.40.0
python-dotenv>=1.0.0

# Data
yfinance>=0.2.40
pandas>=2.0.0
numpy>=1.24.0
requests>=2.31.0

# Technical Indicators (pure Python, no C compilation needed)
pandas-ta>=0.3.14b

# Machine Learning
xgboost>=2.0.0
lightgbm>=4.0.0
scikit-learn>=1.3.0
scipy>=1.11.0

# Dashboard
streamlit>=1.30.0
plotly>=5.18.0
altair>=5.0.0

# Utilities
tqdm>=4.66.0
colorama>=0.4.6
